feat: enhance database configuration and management; add NativePHP environment variables, improve validation, and update UI components

+++ database/migrations/2025_08_25_003044_create_item_sale_details_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->bigInteger('item_sale_id')->index();
             $table->integer('item_id');
             $table->decimal('jumlah')->default(0);
-            $table->decimal('harga')->default(0);
+            $table->double('harga')->default(0);
             $table->integer('unit_id');
         });
     }
